api/requirements.txt
Flask==2.2.3
torch==2.0.1
torchvision==0.15.2
Pillow==9.5.0
requests==2.31.0
Flask-SQLAlchemy==3.0.0
